Fix shadow bug with findOneAndUpdate()
Hello! 

Spend a good part of today trying to figure out why this plugin didn't work and I ended stepping through the entire plugin. Because the variables used for lodash's `each` are `value, key`, value overwrites the upper scope's actual `value` to undefined. So when the query is actually made, the count is always 0 because it's looking for an `email` (or other unique field) that is set to `undefined`